Moved some of the DB code for RVfile into its class, started working on a File Type header detector, started a new file scanning engine that will scan Zip as a file and as an archive, and recognizes files with headers and CHD, and will working with Zip's inside of Zip's ..... more DB speed improvements made for Zero byte files.

+++ RomVaultX/romScanner.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -41,6 +42,132 @@ namespace RomVaultX
             Program.SyncCont = null;
         }
 
+        private static void ScanAFile(FileInfo f)
+        {
+            Debug.WriteLine(f.FullName);
+            
+            Stream fStream;
+            int errorCode = IO.FileStream.OpenFileRead(f.FullName, out fStream);
+            if (errorCode != 0)
+                return;
+
+            int offset;
+            FileType foundFileType = FileHeaderReader.GetType(fStream, out offset);
+
+            RvFile tFile = UnCompFiles.CheckSumRead(fStream, offset);
+            tFile.AltType = foundFileType;
+
+            if (foundFileType == FileType.CHD)
+            {
+                // need to validate check the CHD file
+            }
+
+            // test if needed.
+            FindStatus res = fileneededTest(tFile);
+
+            if (res == FindStatus.FileNeededInArchive)
+            {
+                Debug.WriteLine("Reading file as "+tFile.SHA1);
+                GZip gz = new GZip();
+                gz.crc = tFile.CRC;
+                gz.md5Hash = tFile.MD5;
+                gz.sha1Hash = tFile.SHA1;
+                gz.uncompressedSize = tFile.Size;
+
+                Stream ds;
+                IO.FileStream.OpenFileRead(f.FullName, out ds);
+                string outfile = Getfilename(tFile.SHA1);
+                gz.WriteGZip(outfile, ds, false);
+                ds.Close();
+                ds.Dispose();
+
+                tFile.DBWrite();
+            }
+
+            if (foundFileType == FileType.ZIP)
+            {
+                ZipFile fz = new ZipFile();
+                fz.ZipFileOpen(f.FullName, f.LastWriteTime, true);
+                //if (fz.ZipStatus == ZipStatus.TrrntZip)
+                //{
+
+                //}
+                //else
+                {
+                    for (int i = 0; i < fz.LocalFilesCount(); i++)
+                    {
+                        int Buffersize = 1024 * 1024;
+                        byte[] _buffer = new byte[Buffersize];
+                        Stream stream;
+                        ulong streamSize;
+                        ushort compressionMethod;
+                        fz.ZipFileOpenReadStream(i, false, out stream, out streamSize, out compressionMethod);
+                        string file = @"C:\RomVaultX\" + Guid.NewGuid();
+                        Stream Fs;
+                        IO.FileStream.OpenFileWrite(file, out Fs);
+                        ulong sizetogo = streamSize;
+                        while (sizetogo > 0)
+                        {
+                            int sizenow = sizetogo > (ulong)Buffersize ? Buffersize : (int)sizetogo;
+                            stream.Read(_buffer, 0, sizenow);
+                            Fs.Write(_buffer, 0, sizenow);
+                            sizetogo -= (ulong)sizenow;
+                        }
+                        Fs.Close();
+                        stream.Close();
+
+                        FileInfo fi = new FileInfo(file);
+                        ScanAFile(fi);
+                        File.Delete(file);
+                    }
+                }
+            }
+            if (foundFileType == FileType.GZ)
+            {
+
+            }
+
+        }
+
+        private static void ScanADirNew(string directory)
+        {
+            _bgw.ReportProgress(0, new bgwText("Scanning Dir : " + directory));
+            DirectoryInfo di = new DirectoryInfo(directory);
+
+            FileInfo[] fi = di.GetFiles();
+
+            _bgw.ReportProgress(0, new bgwRange2Visible(true));
+            _bgw.ReportProgress(0, new bgwSetRange2(fi.Count()));
+
+            for (int j = 0; j < fi.Count(); j++)
+            {
+                if (_bgw.CancellationPending)
+                    return;
+
+                FileInfo f = fi[j];
+                _bgw.ReportProgress(0, new bgwValue2(j));
+                _bgw.ReportProgress(0, new bgwText2(f.Name));
+
+                ScanAFile(f);
+            }
+
+            DirectoryInfo[] childdi = di.GetDirectories();
+            foreach (DirectoryInfo d in childdi)
+            {
+                if (_bgw.CancellationPending)
+                    return;
+                ScanADirNew(d.FullName);
+            }
+
+            if (directory == "ToSort")
+                return;
+            if (IsDirectoryEmpty(directory))
+                Directory.Delete(directory);
+
+        }
+
+
+
 
         private static void ScanADir(string directory)
         {
@@ -74,7 +201,7 @@ namespace RomVaultX
                     for (int i = 0; i < fz.LocalFilesCount(); i++)
                     {
                         Debug.WriteLine(fz.Filename(i));
-                        rvFile tFile = new rvFile();
+                        RvFile tFile = new RvFile();
                         tFile.Size = fz.UncompressedSize(i);
                         tFile.CRC = fz.CRC32(i);
                         tFile.MD5 = fz.MD5(i);
@@ -109,7 +236,7 @@ namespace RomVaultX
                             gz.WriteGZip(outfile, zds, isZipTrrntzip);
                             fz.ZipFileCloseReadStream();
 
-                            DataAccessLayer.AddInFiles(tFile);
+                            tFile.DBWrite();
                         }
                     }
                     fz.ZipFileClose();
@@ -120,17 +247,17 @@ namespace RomVaultX
                 }
                 else if (ext.ToLower() == ".gz")
                 {
-                    GZip gZipTest=new GZip();
-                    ZipReturn errorcode=gZipTest.ReadGZip(f.FullName, true);
+                    GZip gZipTest = new GZip();
+                    ZipReturn errorcode = gZipTest.ReadGZip(f.FullName, true);
                     if (errorcode != ZipReturn.ZipGood)
                         continue;
-                    rvFile tFile = new rvFile();
+                    RvFile tFile = new RvFile();
                     tFile.CRC = gZipTest.crc;
                     tFile.MD5 = gZipTest.md5Hash;
                     tFile.SHA1 = gZipTest.sha1Hash;
                     tFile.Size = gZipTest.uncompressedSize;
                     tFile.CompressedSize = gZipTest.compressedSize;
-                    
+
                     FindStatus res = fileneededTest(tFile);
 
                     if (res == FindStatus.FileUnknown)
@@ -152,17 +279,17 @@ namespace RomVaultX
                         ds.Dispose();
 
                         gZipTest.Close();
-                        DataAccessLayer.AddInFiles(tFile);
+                        tFile.DBWrite();
                     }
 
                     File.Delete(f.FullName);
-                    
+
                 }
                 else
                 {
 
 
-                    rvFile tFile = new rvFile();
+                    RvFile tFile = new RvFile();
                     int errorcode = UnCompFiles.CheckSumRead(f.FullName, true, out tFile.CRC, out tFile.MD5, out tFile.SHA1, out tFile.Size);
 
                     if (errorcode != 0)
@@ -189,14 +316,14 @@ namespace RomVaultX
                         ds.Close();
                         ds.Dispose();
 
-                        DataAccessLayer.AddInFiles(tFile);
+                        tFile.DBWrite();
                     }
 
                     File.Delete(f.FullName);
                 }
 
             }
-            
+
             DirectoryInfo[] childdi = di.GetDirectories();
             foreach (DirectoryInfo d in childdi)
             {
@@ -205,6 +332,8 @@ namespace RomVaultX
                 ScanADir(d.FullName);
             }
 
+            if (directory == "ToSort")
+                return;
             if (IsDirectoryEmpty(directory))
                 Directory.Delete(directory);
         }
@@ -228,7 +357,7 @@ namespace RomVaultX
             FoundFileInArchive,
             FileNeededInArchive,
         };
-        private static FindStatus fileneededTest(rvFile tFile)
+        private static FindStatus fileneededTest(RvFile tFile)
         {
             // first check to see if we already have it in the file table
             bool inFileDB = FindInFiles.Execute(tFile); // returns true if found in File table
